+++ src/OrderedListTest.php
@@ -157,24 +157,6 @@ final class OrderedListTest extends StructuredFieldTest
     }
 
     /** @test */
-    public function it_can_sanitize_the_maps(): void
-    {
-        $structuredField = OrderedList::from();
-        $structuredField[] = 42;
-        $structuredField[] = 42;
-        $structuredField[] = 42;
-        unset($structuredField[1]);
-        $item = $structuredField[0];
-        $item->parameters->append('forty-two', '42');
-        $wrongUpdatedItem = $item->parameters->get('forty-two');
-        $wrongUpdatedItem->parameters->append('invalid-value', 'not-valid');
-        self::assertCount(1, $wrongUpdatedItem->parameters);
-
-        $structuredField->sanitize();
-        self::assertCount(0, $wrongUpdatedItem->parameters);
-    }
-
-    /** @test */
     public function it_fails_http_conversion_with_invalid_parameters(): void
     {
         $this->expectException(StructuredFieldError::class);